requirements.txt
crewai
crewai-tools
streamlit
google-generativeai
python-dotenv
